// packages/react-md-map/src/types.ts
// Define types for regions
export type RaionCode = 'CA' | 'GA' | 'TA' | 'CT' | 'BS' | 'LE' | 'CM' | 'SV' | 'BD' | 'CS' | 'HI'
                      | 'IA' | 'NI' | 'CU' | 'AN' | 'CR' | 'ST' | 'UN' | 'DU' | 'CL' | 'SN' | 'FA'
                      | 'OR' | 'GL' | 'BA' | 'TE' | 'SI' | 'RI' | 'RE' | 'SD' | 'FL' | 'DR' | 'ED'
                      | 'BR' | 'SO' | 'DO' | 'OC';